backend/entrypointのエラー解消

+++ backend/entrypoint.sh
@@ -8,7 +8,9 @@ rm $pidfile
 
 port=18101
 pre_process="$(ps -ef |grep ':$port' |grep -v 'grep' |awk '{print $2}')"
-kill -9 $pre_process
+if [ -n "$pre_process" ]; then
+  kill -9 $pre_process
+fi
 # RAILS_ENV=${RAILS_ENV} ./bin/rails assets:precompile # M1Macで動かなかったから、graphiql以外のviewをいじるまでコメントアウト
 RAILS_ENV=${RAILS_ENV} bundle exec rails s -p $port -b "0.0.0.0" &
